Add files via upload
The objective of this project is to perform exploratory data analysis, clean and correct, prepare the data, test various machine learning techniques, and select the best model based on the performance. Another critical product of this study will be to record the accuracy of the predictions and which method improves the predictability.